refactor: move services to dedicated modules
Closes #53
