+++ test/unit/services/receipt/test_service.py
@@ -2,7 +2,12 @@ from unittest.mock import Mock, sentinel
 
 import pytest
 
-from taxes.services import receipt
+from taxes.services.receipt.service import (
+    create as create_receipt_service,
+    Dependency as ReceiptServiceDependency,
+    ReceiptService,
+)
+from taxes.services.receipt.entities.receipt import ItemToInsert
 from receipt.use_cases.test_create_receipt import create_receipt_test_cases
 
 
@@ -10,24 +15,24 @@ from receipt.use_cases.test_create_receipt import create_receipt_test_cases
 def make_dependencies_fixture():
     def build():
         return {
-            'logger': Mock(spec=receipt.Dependency.Logger),
-            'basket_service': Mock(spec=receipt.Dependency.BasketService),
-            'tax_service': Mock(spec=receipt.Dependency.TaxService),
+            'logger': Mock(spec=ReceiptServiceDependency.Logger),
+            'basket_service': Mock(spec=ReceiptServiceDependency.BasketService),
+            'tax_service': Mock(spec=ReceiptServiceDependency.TaxService),
         }
     return build
 
 
 def test_create_returns_service(make_dependencies_fixture):
-    service = receipt.create(**make_dependencies_fixture())
-    assert isinstance(service, receipt.ReceiptService)
+    service = create_receipt_service(**make_dependencies_fixture())
+    assert isinstance(service, ReceiptService)
 
 
 @create_receipt_test_cases
 def test_create_receipt_returns_receipt(case, make_dependencies_fixture):
-    service = receipt.create(**make_dependencies_fixture())
+    service = create_receipt_service(**make_dependencies_fixture())
     service.basket_service.create_basket.side_effect = lambda a: a
     service.tax_service.add_taxes.return_value = [
-        receipt.ItemToInsert(
+        ItemToInsert(
             description=p.article_in_basket.product.name,
             quantity=p.article_in_basket.quantity,
             unit_price_before_taxes=p.article_in_basket.product.unit_price,
